refactor: Implement data-driven skills and talents
This commit introduces a more robust data-driven system for handling player skills and talents, and migrates several abilities to this new system.

Key changes:
- Created a new `skillProcessor.ts` to centralize skill effect logic, refactoring it out of the main `gameStore.ts`.
- Implemented rank-based scaling for skills. The `skillProcessor` can now correctly parse values from arrays in the skill's data definition, allowing effects to scale with player investment.
- Migrated 'Mage - Fireball' and 'Berserker - Heroic Strike' to be fully data-driven with rank scaling.
- Implemented two previously non-functional talents:
    - 'Fracas d'Armure' (passive armor ignore on hit).
    - 'Sanguinaire' (triggered bleed effect on critical strike).
- Improved the combat log by adding a new `talent_proc` message type with a distinct color for better player feedback.
- Added extensive comments in `gameStore.ts` and `skillProcessor.ts` to document the large scope of remaining unimplemented features (especially the need to convert all skills and the lack of a monster ability system), providing a clear path for future development.
